Update the build script. The YML file is now optional. If it not defined, the whole YML repo will be added to the docker image; otherwise only the specific file will be added, and it will be pass as an startup argument to the server, as well as the flag to disable the HW auto-detection

+++ scripts/build.sh
@@ -41,23 +41,33 @@ fi
 if [ -z ${yml_use_local+x} ]; then
     echo "Getting yml file from ${yml_repo}"
 
-    # This repository doesn't use assent, so we need to clone the repository
-    # and copy the file we want
-    git clone ${yml_repo} -b ${yml_repo_tag} yml_repo || exit 1
-    mv yml_repo/defaults/${yml_file_name} local_files || exit 1
-    rm -rf yml_repo
+    if [ -z ${yml_file_name} ]; then
+        # If an specific file was not define, clone the whole repository
+        git -C local_files clone ${yml_repo} -b ${yml_repo_tag} || exit 1
+    else
+        # If an specific file was defined, copy it
+        git clone ${yml_repo} -b ${yml_repo_tag} yml_repo || exit 1
+        mv yml_repo/defaults/${yml_file_name} local_files || exit 1
+        rm -rf yml_repo
+
+        # Additionally, when an specific YML file is defined, add the '--disable-hw-detect'
+        # option to the list of server arguments as well as the specified YML file using the
+        # '-d' option.
+        server_args+=" --disable-hw-detect -d /tmp/fw/${yml_file_name}"
+    fi
+
 else
     echo "Using local yml file..."
 fi
 
-# Divide the server argument string into a list of quoted substring, divided by comas.
+# Remove any white spaces and the beginning or end of the server argument string.
+# Then divide it into a list of quoted substring, divided by comas.
 # This is the format that the Dockerfile uses
-server_args_list=$(echo \"${server_args}\" | sed 's/\s/","/g')
+server_args_list=$(echo \"${server_args}\" | sed 's/^\s*//g'| sed 's/\s*$//g' | sed 's/\s/","/g')
 
 # Generate the Dockerfile from the template
 cat Dockerfile.template \
         | sed s/%%PYSMURF_SERVER_BASE_VERSION%%/${pysmurf_server_base_version}/g \
-        | sed s/%%YML_FILE_NAME%%/${yml_file_name}/g \
         | sed s/%%SERVER_ARGS%%/"${server_args_list}"/g \
         > Dockerfile
